+++ CHANGELOG.md
@@ -3,9 +3,21 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
- 
-## [1.0.0] - 2025-09-XX
+
+## Unreleased
+
+### Changed
+- Updated FAQs to reflect switch to LSTM <50 model, trained on observed streamflow values below the 50th percentile.
+- Dropped 'early release' language from title dialog, page banner, and metadata
+
+## [1.1.0] - 2026-01-27
+
+### Changed
+
+- Update name of application to "River DroughtCast — Streamflow drought status and forecasts"
+
+## [1.0.0] - 2025-12-10
  
 ### Added
 
-- Initial public release of the Streamflow drought assessment and forecasting tool
+- Initial early public release of the Streamflow drought assessment and forecasting tool